refactor: move webpack config files #wip
