+++ webpack.config.base.js
@@ -3,7 +3,7 @@
  */
 
 const path = require("path");
-const { dependencies: externals } = require("../app/package.json");
+const { dependencies: externals } = require("./app/package.json");
 
 module.exports = {
   module: {
