+++ webpack.config.production.js
@@ -15,8 +15,8 @@ module.exports = merge(baseConfig, {
   entry: ["./app/index"],
 
   output: {
-    path: path.join(__dirname, "./app/dist"),
-    publicPath: "./dist/"
+    path: path.join(__dirname, "app/dist"),
+    publicPath: "../dist/"
   },
 
   module: {
@@ -92,7 +92,7 @@ module.exports = merge(baseConfig, {
 
     new HtmlWebpackPlugin({
       filename: "../app.html",
-      template: "../app/app.html",
+      template: "app/app.html",
       inject: false
     })
   ],
